# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polymarket Bot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets:
    get:
      operationId: listMarkets
      tags: [markets]
      summary: List watched markets
      description: Returns all markets in the watchlist ordered by score
      responses:
        "200":
          description: Market list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListResponse"
    post:
      operationId: upsertMarkets
      tags: [markets]
      summary: Upsert markets
      description: Insert or update markets from a fresh watchlist scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                markets:
                  type: array
                  items:
                    $ref: "#/components/schemas/MarketInput"
              required: [markets]
      responses:
        "200":
          description: Upsert result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpsertResult"
...
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List recent signals
      description: Returns recent signals from all strategy engines
      parameters:
        - name: strategy
          in: query
          schema:
            type: string
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Signal list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalListResponse"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create a signal
      description: Insert a new strategy signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalInput"
      responses:
        "201":
          description: Created signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalCreated"
...
components:
  schemas:
...
    Market:
      type: object
      properties:
        marketId:
          type: string
        conditionId:
          type: string
          nullable: true
        question:
          type: string
          nullable: true
        eventTitle:
          type: string
          nullable: true
        eventSlug:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        negRisk:
          type: boolean
          nullable: true
        tokenIds:
          type: array
          items:
            type: string
          nullable: true
        outcomes:
          type: array
          items:
            type: string
          nullable: true
        volume24h:
          type: string
          nullable: true
        liquidity:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        hoursToClose:
          type: string
          nullable: true
        feesEnabled:
          type: boolean
          nullable: true
        score:
          type: string
          nullable: true
        addedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - marketId
    MarketInput:
      type: object
      properties:
        marketId:
          type: string
        conditionId:
          type: string
          nullable: true
        question:
          type: string
          nullable: true
        eventTitle:
          type: string
          nullable: true
        eventSlug:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        negRisk:
          type: boolean
          nullable: true
        tokenIds:
          type: array
          items:
            type: string
          nullable: true
        outcomes:
          type: array
          items:
            type: string
          nullable: true
        volume24h:
          type: string
          nullable: true
        liquidity:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        hoursToClose:
          type: string
          nullable: true
        feesEnabled:
          type: boolean
          nullable: true
        score:
          type: string
          nullable: true
      required:
        - marketId
    MarketListResponse:
      type: object
      properties:
        markets:
          type: array
          items:
            $ref: "#/components/schemas/Market"
        count:
          type: integer
      required:
        - markets
        - count
    UpsertResult:
      type: object
      properties:
        upserted:
          type: integer
      required:
        - upserted
...
    Signal:
      type: object
      properties:
        id:
          type: integer
        strategy:
          type: string
        marketId:
          type: string
          nullable: true
        eventSlug:
          type: string
          nullable: true
        signalScore:
          type: string
          nullable: true
        metadata:
          nullable: true
        emittedAt:
          type: string
          format: date-time
          nullable: true
        entryPrice:
          type: string
          nullable: true
        exitPrice:
          type: string
          nullable: true
        pnl:
          type: string
          nullable: true
        resolved:
          type: boolean
          nullable: true
        question:
          type: string
          nullable: true
        executed:
          type: boolean
          nullable: true
          description: >
            True once the executor has handled this signal (placed an order
            or skipped with a reason). Prevents re-processing each cycle.
        executedSkipReason:
          type: string
          nullable: true
          description: >
            Set when the executor skipped placing an order for this signal
            (e.g. size_below_floor, missing_token_id). Mutually exclusive
            with orders.status='REJECTED' which records CLOB-side rejections.
      required:
        - id
        - strategy
    SignalInput:
      type: object
      properties:
        strategy:
          type: string
        marketId:
          type: string
          nullable: true
        eventSlug:
          type: string
          nullable: true
        signalScore:
          type: string
          nullable: true
        metadata:
          nullable: true
        entryPrice:
          type: string
          nullable: true
        exitPrice:
          type: string
          nullable: true
        pnl:
          type: string
          nullable: true
        resolved:
          type: boolean
          nullable: true
      required:
        - strategy
    SignalCreated:
      type: object
      properties:
        id:
          type: integer
          description: "-1 if deduplicated (same signal within the last hour)"
      required:
        - id
    SignalListResponse:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        count:
          type: integer
      required:
        - signals
        - count
...